[dnssd-platform] check `OTBR_ENABLE_DNSSD_PLAT` config
This commit adds `#if` checks for `OTBR_ENABLE_DNSSD_PLAT` to the
`dnssd` modules, ensuring the `otPlatDnssd` APIs are not provided
unless this feature is enabled.

The `OTBR_ENABLE_DNSSD_PLAT` option controls the corresponding
OpenThread core `OT_PLATFORM_DNSSD` configuration, which indicates
that `otPlatDnssd` APIs are provided by the platform layer.