[config] set `OT_BORDER_AGENT_MESHCOP_SERVICE` based on OTBR config
This commit updates `openthread/CMakeLists.txt` to ensure that
`OT_BORDER_AGENT_MESHCOP_SERVICE` (which controls whether OpenThread
core's Border Agent registers the MeshCoP service) is set to the inverse
of `OTBR_BORDER_AGENT_MESHCOP_SERVICE`. This ensures that either
OpenThread core or the OTBR platform handles the MeshCoP service
registration and the configs are consistent.