[border-agent] remove "rv" key from vendor TXT data (#2862)
This commit updates `EncodeVendorTxtData()` to no longer include
the "rv" key. This is because the "rv" key is now included as part of
the TXT data generated by OpenThread core's Border Agent.

The "rv" key represents the version of the TXT record format, and per
the Thread specification, it must be set to "1".