[ci] add raspbian build check for bookworm version (#2895)
This commit extends the test scripts so that it can test with the
recently raspbian OS versions. The recent raspbian OS can only be
downloaded in `.xz` format instead of `.zip` so the current test
script doesn't work with it. The commit also adds a test to verify the
latest 2025-05-13 version of raspios.