[bootstrap] improve robustness of mDNSResponder download and extraction (#2877)
The mDNSResponder download and build process was susceptible to
several issues:

1. Incomplete or Corrupted Downloads: If the `wget` command failed to
download the entire mDNSResponder archive, or if the download was
corrupted, the script might continue with an invalid archive.

2. Pre-existing Files: If the `/tmp/mDNSResponder-2600.100.147`
directory or `mDNSResponder.tar.gz` file already existed from a
previous run, the `tar` command would fail to extract the archive
correctly, leading to a mix of old and new files.

The commit addresses the two issues with:
* The `tar --overwrite ` command is to overwrite existing files when
  extracting.
* The `wget -O mDNSResponder.tar.gz` command is used to explicitly
  specify the output file name and overwrite the old one.