[border-agent] add "vendor TXT Data changed" callback (#2941)
This commit adds a callback to `otbr::BorderAgent` to signal changes
to the vendor-specific TXT data.

This callback is used when `OTBR_ENABLE_BORDER_AGENT_MESHCOP_SERVICE`
is disabled, meaning the OpenThread core is responsible for
advertising the Border Agent's MeshCoP service. The new mechanism
allows `otbr::BorderAgent` to notify the `ThreadHost`(e.g., `RcpHost`
or `NcpHost`) to then pass the updated vendor TXT data to the
OpenThread stack.

To support this, a new virtual method `SetBorderAgentVendorTxtData()`
is added to the `ThreadHost` interface. An implementation is provided
for `RcpHost`, with an empty implementation for `NcpHost` to be added
in the future.