[common] move include of byteswap.hpp into source (#2977)
Recently I met a build issue on CI on MacOS (with my code change). The
cause is that the protobuf header and `byteswap.hpp` has
re-definition. `types.hpp` is widely included by sources in
`otbr-agent` while `byteswap.hpp` is only used by some implementations
of `types.hpp`. So we can move the usage of `byteswap.hpp` into
`types.cpp` to solve this issue.