[mDNS] fix avahi issue (#2998)
This commit fixes the wrong MDNS options in CI for NCP and 2 bugs in
MdnsAvahi.

In the MDNS matrix in CI for NCP, when the MDNS is avahi, the otbr
docker image actually still ran with the default MDNS option because
the docker image was built in the CI by command so the existing
environment variable in CI doesn't affect it. I added a `build-arg` to
set the MDNS option for the docker image.

Avahi client has a registering state. Publishing service during this
state leads to failure. But currently `IsStarted` method also regards
this state as `true`. That's why we failed in the NCP test
case. Because the meshcop service was published during this state.

Another issue is that in NCP mode, the BorderAgent isn't added as an
observer to the mDNS. mDNSResponder can work because its state becomes
ready in an earlier stage. But for avahi whose state becomes ready
later than the meshcop service initial publishing, the issue happens.