[mdns] fix ServiceInstanceResolutions leak in mdnssd
ServiceInstanceResolution objects transiently stored in
PublisherMDnsSd::ServiceSubscription::mResolvingInstances
were not being removed after completion of the resolve operation.
This fix removes the object from the mResolvingInstances vector
and keeps it in scope long enough for callbacks to safely
reference. This leak was preventing DNSServiceRefs allocated
for AddrInfo resolution to be deallocated, thus leaking
connections (and related fds) to mdnsResponder's domain socket.