[tests] enhance flaky test netif unit test (#3000)
This commit enhances the flaky test case
`Netif.WpanIfRecvIp6PacketCorrectly_AfterReceivingFromNetif`.

This commit updates the consumer side of the test to use a select
instead of a blocking recv in a Thread. In addition, the PR puts the
producer side of the data packet into a thread and delay the producing
of it.

This seems to be effective. I tested on github CI using tmate with
this change and I can pass the test for 20 times continuously. As a
comparison, it's easy to get this test failed on github CI instance by
running it about 3 times.