[dnssd] implement Dnssd Browser/Resolver APIs (#2962)
The commit implements all the dnssd platform APIs for browse/resolve
except for `RecordQuerier` because currently the mdns publisher
doesn't support record query. I plan to support this later.

The background for this commit is that I plan to use the natvie OT
discovery proxy and remove the current Discovery Proxy module in
ot-br-posix. I've noticed that OT now even supports native mDNS. But
to support DNSSD runtime selection
(OPENTHREAD_CONFIG_PLATFORM_DNSSD_ALLOW_RUN_TIME_SELECTION), we will
also need an implementation on the posix platform.

The implementation in this commit basically follows the old PR
https://github.com/openthread/ot-br-posix/pull/2042. But there are
some major differences in the dnssd platform API definition since
then. The core idea of the implementation for browse/resolve is that
we maintain a map of query type to a map of netif index to a vector of
callback. Whenever the dnssd gets some result, it returns the result
to all the callbacks that are interested to it.

```
Map_BrowsersTable
|  DnsName  |  Map_BrowserCallbackTable  |
|  DnsName  |  Map_BrowserCallbackTable  |
|  DnsName  |  Map_BrowserCallbackTable  |

Map_BrowserCallbackTable
|  NetifIndex  |   Vector_Callbacks   |
|  NetifIndex  |   Vector_Callbacks   |
|  NetifIndex  |   Vector_Callbacks   |
```

As a contrast, the old implementation doesn't need a map of map
because there is only one callback. But now we have a callback for
each browse/resolve action.