[script] add dbus reload in ncp_mode script setup (#3014)
This commit adds dbus reload step in the test setup.

From the log it turned out that the dbus policy wasn't correctly
configured and this caused the cli command to fail. That's why there
are some randomness here and sometimes the case can pass because the
dbus policy is correctly set.