[mdns] avoid duplicate service instance resolutions
This change prevents scheduling a new service instance resolution if an
identical resolution is already in progress in `PublisherMDnsSd` class
(mDNS `Publisher` based on mDNSResponder).

A new `ServiceInstanceResolution::Matches()` method is added to check
if a resolution request matches an existing one. The matching is based
on interface index, instance name, type, and domain. The string
comparisons are case-insensitive.

`ServiceSubscription::Resolve()` is updated to iterate through the list
of active resolutions and use the new `Matches()` method to check for
duplicates before creating/starting a new service resolution.