[scripts] introduce systemd-networkd for network management (#3012)
This commit introduces support for `systemd-networkd` to manage the
host's network configuration. This is necessary to support newer OS
versions like Debian 12 on reference devices, where `systemd-networkd`
is a common network manager.

A new script, `script/_systemd_networkd`, is added. It is controlled
by the `SYSTEMD_NETWORKD` build flag.

The script also includes a check for `REFERENCE_DEVICE` as we enable
OTBR's built-in DHCPv6 PD client on reference device. If OTBR's
built-in DHCPv6 PD client is active, it disables `systemd-networkd`'s
client to prevent port conflicts.