[ncp] add DBus interface to set the host power mode (#3028)
The NCP monitors the host's power state using the
SPINEL_PROP_HOST_POWER_STATE property. Previously, the host was unable
to modify this property on the NCP. This commit adds support for
updating the host power state property on the NCP. As a result, the
NCP can now accurately track the host's power state and wake the host
when necessary.