[main] move the body of the `main()` method to the `realmain.cpp` file
The file `main.cpp` includes many functions for the OTBR Agent. To generate
a library that contains all functions of the OTBR Agent, these functions in
`main.cpp` should be re-used in the library.

This commit moves all methods in `main()` to a new file `realmain.cpp` and provdes
an API for running the OTBR Agent.