[trel] fix build dependency (#3039)
This commit fixes the unnecessary dependency of border-agent on
trel-dnssd. BorderAgent doesn't have any dependency on TrelDnssd. This
is caused by a missing header. This commit adds the missing
header. But even we added the header, the TrelDnssd won't be built and
there will be a link issue because it's dependent by agent module
instead of the BorderAgent module.