[github-actions] fix raspbian check (#3052)
This commit replaces `su -c 'RELEASE=1 script/bootstrap' pi` with
`RELEASE=1 script/bootstrap`. The original command specified running
the script as the user 'pi' but the script ran `sudo`. I guess this is
why it failed.

The weird thing is that this have been working for a long time. I
guess the issue happened recently because the qemu package installed
is updated.

This commit updates the CI environment of raspbian from Ubuntu-22.04
to Ubuntu-24.04 by the way.