[dnssd] update the implementation of browse/resolve methods (#3048)
This commit enhances the implementation of dnssd discovery proxy
methods and fixes the issue of starting browse/resolve in the callback
of these methods.

The general idea is:
1. The callback tables will all be updated immediately when doing
   start/stop browse/resolve operations. Unlike the current
   implementation: A delete operation is done in a lazy way. And an
   add operation is simply discarded if invoking callbacks.
2. When invoking callbacks, we will make a copy of all the callbacks
   and iterate over the copy. So the copy won't change during the
   invoke process.
3. To control what's to subscribe and unsubscribe via mPublisher, 3
   std::set members are introduced to track what are currently being
   subscribed. Each time when the callback maps change, a task will
   posted to update the subscriptions based on the difference between
   the callback maps and the set.

This way the implementation has better readability and can support
flexible usages. I have also verified the implementation with
`OT_DNSSD_DISCOVERY_PROXY` and it can work (a ot-cli-ftd use dns
client to discover the meshcop service of a BR).

This change is required to support this because the OT discovery proxy
has some usages of adding new subscriptions in the callback of the
previous dnssd operation.