[ncp] add local port to udp forward callback (#3069)
This commit adds one more parameter `aLocalPort` to
`UdpForwardToHostCallback`.

Currently we only use udp proxy for border agent. As we are going to
support Epskc and TREL, more UDP proxies will be required. We need to
have the local port to find the right UDP proxy to send the UDP
packet.