[dbus-test] remove check for 'nn' (network name) in meshcop service
This commit removes the test steps that verify the presence of the
`nn` (network name) TXT entry in the MeshCoP DNS-SD service.

This change aligns the test with a recent behavior modification in the
OpenThread core (PR 11989), which no longer advertises the `nn`
entry if a valid network dataset is not available. This prevents a
Border Agent from prematurely advertising a default or invalid
Network Name or Extended PAN ID during boot-up before it gets valid
network information.