[test] remove check for `nn` and `xp` in `ncp_border_agent.exp`
This commit removes the test steps that verify the presence of the
`nn` (network name) and `xp` (Extended PAN ID) TXT entry in the
`ncp_border_agent.exp`

This change aligns the test with a recent behavior modification in the
OpenThread core, which no longer advertises the `nn` entry if a
valid network dataset is not available. This prevents a Border Agent
from prematurely advertising a default or invalid Network Name or
Extended PAN ID during boot-up before it gets valid network
information.