[script] enable mDNSResponder for OTBR reference-device for testing
This enables mDNSResponder with its cli tool 'dns-sd' which is used for Thread cert tests.
The service is not started by default, to not interfere with the existing OT mDNS which
is active by default. That means the tests using 'dns-sd' must manually start the
service. Also, Avahi is disabled for the OTBR reference-device as it's not used and it
interferes with proper host naming: it was observed that Avahi causes the hostname to
frequently change with an ever-increasing number.