[discovery-proxy] use OT dnssd proxy for system mDNS cases (#3099)
This commit changes the DiscoveryProxy implementation to OT
implementation in all cases.

Currently OT Discovery Proxy is used when `OTBR_MDNS` is
openthread. This PR makes OT Discovery Proxy the default option in
both cases (`OTBR_MDNS` is openthread or avahi, mDNSResponder). When
`OTBR_MDNS` is not openthread, the dnssd platform implementations that
are completed recently will be used.

I made some basic verification that the discovery poxy can work with OT
implementation. We still keep the option to use the Discovery Proxy in
ot-br-posix in case there are some regressions. Later we can remove
the Discovery Proxy in ot-br-posix.