[ncp] support dns discovery proxy for service browser (#3141)
This commit adds support for Dnssd discovery proxy in NCP on host
side.

The commit handles `SPINEL_PROP_DNSSD_BROWSER`, utilizes
`DnssdPlatform` on host side to do the browse, and sends back the
result to NCP through spinel.

The commit also subscribes to the platform dnssd in NCP so that it can
get notified when service/host are discovered.